+++ CHANGES.rst
@@ -29,11 +29,18 @@ Version 3.1.0 (2019-XX-XX)
 
 - fix issue where unreferenced files were not being removed when deleting an entire dataset
 
+- support nested :class:`~kartothek.io_components.metapartition.MetaPartition`
+  in :meth:`~kartothek.io_components.metapartition.MetaPartition.add_metapartition`.
+  This fixes issue https://github.com/JDASoftwareGroup/kartothek/issues/40 .
+
+- Add :meth:`~karothek.io.dask.bag.build_dataset_indices__bag`
+
 **Breaking:**
 
 - categorical normalization was moved from :meth:`~kartothek.core.common_metadata.make_meta` to
   :meth:`~kartothek.core.common_metadata.normalize_type`.
 - :meth:`kartothek.core.common_metadata.SchemaWrapper.origin` is now a set of of strings instead of a single string
+- ``Partition.from_v2_dict`` was removed, use :meth:`kartothek.core.partition.Partition.from_dict` instead
 
 
 Version 3.0.0 (2019-05-02)